fast-api/requirements.txt
fastapi
uvicorn
openai
python-dotenv
passlib[bcrypt]
pyjwt

